UI Reorganisation and new layout concept started
