+++ interface/src/application/pages/DevicePages/ControlPalette/ManualJogPalette.tsx
