+++ interface/src/application/pages/DevicePages/ControlPalette/SceneControl/state.md
