+++ interface/src/application/pages/DevicePages/ControlPalette/TrackPalette.tsx
