+++ interface/src/application/pages/DevicePages/DataViews/ThreeD/PolyLineCurve.tsx
