+++ interface/src/application/pages/DevicePages/archive/SystemPage.tsx
@@ -27,8 +27,8 @@ import { Composition, Box } from 'atomic-layout'
 import { Printer } from '@electricui/components-desktop'
 import { RouteComponentProps } from '@reach/router'
 
-import LEDCalibrationModal from '../../components/LEDCalibrationModal'
-import KinematicsDiagram from '../../components/KinematicsDiagram'
+import LEDCalibrationModal from '../../../components/LEDCalibrationModal'
+import KinematicsDiagram from '../../../components/KinematicsDiagram'
 
 const SensorsActive = () => {
   const sensorEnabledState =
