+++ interface/src/application/pages/DevicePages/index.tsx
@@ -1,17 +1,14 @@
 import './device-pages.css'
 
 import { RouteComponentProps, Router } from '@reach/router'
-
-import ControlPage from './ControlPage'
-import DetailPage from './DetailPage'
-import { DisconnectionModal } from '@electricui/components-desktop-blueprint'
-import Footer from '../../components/Footer'
-import Header from '../../components/Header'
-import { Intent } from '@blueprintjs/core'
 import React from 'react'
-import { SceneController } from './SceneControl/SceneController'
-import SystemPage from './SystemPage'
+
+import { DisconnectionModal } from '@electricui/components-desktop-blueprint'
 import { navigate } from '@electricui/utility-electron'
+import { Intent } from '@blueprintjs/core'
+
+import Header from '../../components/Header'
+import MainWindow from './MainWindow'
 
 interface InjectDeviceIDFromLocation {
   deviceID?: string
@@ -40,13 +37,8 @@ const DevicePages = (
       <div className="device-pages">
         <Header deviceID={props.deviceID} {...props} />
         <div className="device-content">
-          <Router>
-            <ControlPage path="/" />
-            <SystemPage path="system" />
-            <DetailPage path="detail" />
-          </Router>
+          <MainWindow />
         </div>
-        <Footer deviceID={props.deviceID} {...props} />
       </div>
     </React.Fragment>
   )
